requirements.txt
pandas
matplotlib
prophet
statsmodels
python-docx